TGP-2889: Navigation bug fix
